Remove member for creating new datasets from document
