+++ veusz/datasets/__init__.py
@@ -27,3 +27,4 @@ from .expression import *
 from .plugin import *
 
 from .commonfn import *
+from .helpers import *
