+++ veusz/datasets/twod.py
@@ -17,6 +17,9 @@
 #    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 ###############################################################################
 
+"""Two dimensional datasets."""
+
+from __future__ import division
 import numpy as N
 
 from ..compat import crepr, cstr
